Fixed the header

+++ src/cstring.h
@@ -8,4 +8,4 @@ size_t cstrlen(const char* str);
 
 int cstrcmp(const char* lhs, const char* rhs);
 
-void cint_to_string(ssize_t input, char* string_ret, size_t ret_size);
+void cint_to_string(int64_t input, char* string_ret, size_t ret_size);
